Undefined Error And Get Settings Flag (#91)
* Fixing undefined error and get settings flag

Fixes undefined error when first initialize/clean the frontend and makes the get settings flag more robust by placing it in the actual get settings file.

* Converting to Modes "enum"

Data returned in cameraGimbalOperations.js on a successful settings update had the mode as a string. This has been changed so that the mode should be a value from the Modes "enum." This should now match the cameraGimbalSettings.js file, and the grayed-out save button issue should be fixed (was previously not gray when it should have been). Also fixed minor issue in airdropSettings.js.

* Updating mode conversion

Switching from operations file to reducer file for cameraGimbal

* Cleaning code

Some organizational and other small improvements.

* Fixing newMode

Simplifies code